[Fix] 수정 내역
1. 다이아 UI 재활성화
2. 코드 최적화 2차
3. 광고 오류 문제 AI 활용 문제 지점 수정
4. 음식, 손님 이미지 최적화

+++ Assets/Scripts/Manager/StaffDataManager.cs
@@ -58,7 +58,7 @@ public class StaffDataManager : MonoBehaviour
     public List<StaffData> GetStaffDataList(EquipStaffType type)
     {
         int typeIndex = (int)GetStaffGroupType(type);
-        return _staffTypeDataList[(int)typeIndex];
+        return _staffTypeDataList[typeIndex];
     }
 
     public List<StaffData> GetSortStaffDataList(EquipStaffType type)
